refactor: extract environment_callback into minarch_environment.c
Move set_rumble_state and environment_callback (the libretro environment
protocol handler) out of the monolithic minarch.c into a dedicated module.
Config_readOptions is made non-static so the environment module can call it
when the core sets its options at init time.
