+++ workspace/all/minarch/minarch_environment.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool environment_callback(unsigned cmd, void *data);
